S104 split: extract ngfw_runtime.py
